+++ lib/cloud-metadata/aws.js
@@ -2,7 +2,7 @@
 const net = require('net')
 const https = require('https')
 const http = require('http')
-const { elRequest } = require('../../elrequest')
+const { elRequest } = require('../elrequest')
 
 /**
  * Checks for metadata server then fetches data
