+++ lib/cloud-metadata/index.js
