+++ test/cloud-metadata/_fixtures.js
