+++ test/cloud-metadata/_lib.js
