+++ test/cloud-metadata/test-server.js
